refactor src ai
