+++ ai/model/src/agents/plan_model.py
@@ -20,8 +20,8 @@ from datetime import date, datetime, timedelta
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-from .promts.plan_promt import JSON_SCHEMA_EXAMPLE, system_plan_prompt
-from .utils.helper_function import extract_image_url
+from ..promts.plan_promt import JSON_SCHEMA_EXAMPLE, system_plan_prompt
+from ..utils.helper_function import extract_image_url
 
 from dotenv import load_dotenv
 from langchain_core.output_parsers import JsonOutputParser
