+++ ai/model/src/models/__init__.py
@@ -1,0 +1,38 @@
+from .accommodation_models import (
+    ImageInfo,
+    RoomType,
+    Accommodation,
+    AccommodationData
+)
+
+from .place_models import (
+    Location,
+    Place,
+    PlaceData
+)
+
+from .restaurant_models import (
+    Restaurant,
+    RestaurantData
+)
+
+from .trip_models import (
+    TripPlanRequest,
+    TripPlanResponse,
+    SimpleTripPlanRequest
+)
+
+__all__ = [
+    'ImageInfo',
+    'RoomType',
+    'Accommodation',
+    'AccommodationData',
+    'Location',
+    'Place',
+    'PlaceData',
+    'Restaurant',
+    'RestaurantData',
+    'TripPlanRequest',
+    'TripPlanResponse',
+    'SimpleTripPlanRequest'
+]
