+++ ai/model/src/models/reponse_models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class SuggestWithIDAndType(BaseModel):
+    name: str
+    type: str
+    args: str
+    id: str
