+++ ai/model/src/promts/__init__.py
@@ -1,0 +1,3 @@
+from .function_promt import query_hotels, query_places, query_fnb, search_by_price_range, search_by_rating, search_by_category, search_by_location, search_by_menu_item
+from .travel_promt import travel_suggestion_system_prompt
+
