+++ ai/model/src/test_api/test_plan_direct.py
@@ -14,7 +14,7 @@ if src_dir not in sys.path:
     sys.path.append(src_dir)
 
 # Import the plan model
-from plan_model import PlanModel
+from ai.model.src.agents.plan_model import PlanModel
 
 def main():
     # Sample input data with multiple locations
