+++ ai/model/src/utils/logger.py
@@ -1,0 +1,54 @@
+import logging
+import os
+from datetime import datetime
+
+# Create logs directory if it doesn't exist
+LOGS_DIR = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'logs')
+os.makedirs(LOGS_DIR, exist_ok=True)
+
+# Configure logging
+def setup_logger(name: str) -> logging.Logger:
+    """
+    Set up and return a logger with consistent configuration
+    
+    Args:
+        name (str): Name of the logger (usually __name__)
+        
+    Returns:
+        logging.Logger: Configured logger instance
+    """
+    # Create logger
+    logger = logging.getLogger(name)
+    logger.setLevel(logging.DEBUG)
+    
+    # Create formatters
+    file_formatter = logging.Formatter(
+        '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+    )
+    console_formatter = logging.Formatter(
+        '%(asctime)s - %(levelname)s - %(message)s'
+    )
+    
+    # Create handlers
+    # File handler - logs everything
+    current_date = datetime.now().strftime('%Y-%m-%d')
+    file_handler = logging.FileHandler(
+        os.path.join(LOGS_DIR, f'app_{current_date}.log'),
+        encoding='utf-8'
+    )
+    file_handler.setLevel(logging.DEBUG)
+    file_handler.setFormatter(file_formatter)
+    
+    # Console handler - only shows INFO and above
+    console_handler = logging.StreamHandler()
+    console_handler.setLevel(logging.INFO)
+    console_handler.setFormatter(console_formatter)
+    
+    # Add handlers to logger
+    logger.addHandler(file_handler)
+    logger.addHandler(console_handler)
+    
+    return logger
+
+# Create default logger for the application
+logger = setup_logger('travel_app')
